Refactored `LampTask.cpp` for cleanliness and minor improvement.
This structure is easier to understand and, while it lightly duplicates
the code, it separates the various cases and makes the execution easier
to understand.

Saving the `millis()` value at the beginning of the task and using it
instead of repeatedly calling `millis()`  allows for the whole run to
behave coherently, avoiding different behaviours as time progresses
during the execution.